+++ docs/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Disable flaw drafts creation for NVD collector (OSIDB-3256)
 - Select most relevant CVE, CVSS, CWE, Source for Vulnerability trackers (OSIDB-3348)
+- Tracker validations show affect's module/component (OSIDB-3439)
 
 ### Fixed
 - Rework and complete the tracker stream pre-selection module to fix it